@@ -2,7 +2,7 @@ name: example
 description: A new Flutter project.
 
 environment:
-  sdk: ">=2.10.0 <3.0.0"
+  sdk: ">=3.4.3 <4.0.0"
   flutter: ">=0.1.4 <2.0.0"
 
 dependencies:
