@@ -4,8 +4,7 @@ version: 2.0.0
 homepage: https://github.com/peng8350/flutter_pulltorefresh
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.0.0"
+  sdk: ">=3.4.3 <4.0.0"
 
 dependencies:
   flutter:
