@@ -4,13 +4,8 @@
  * Time:  2019-08-02 19:20
  */
 
-import 'package:flutter/animation.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
-import 'package:flutter/widgets.dart';
-import 'package:flutter/material.dart'
-    hide RefreshIndicator, RefreshIndicatorState;
-import 'package:pull_to_refresh/src/internals/indicator_wrap.dart';
+import 'package:flutter/material.dart' hide RefreshIndicator, RefreshIndicatorState;
 import 'dart:math' as math;
 import 'package:flutter/physics.dart';
 
@@ -37,18 +32,18 @@ class BezierHeader extends RefreshIndicator {
   // container height(not contain bezier)
   final double rectHeight;
 
-  BezierHeader(
-      {this.child: const Text(""),
-      this.onOffsetChange,
-      this.onModeChange,
-      this.readyRefresh,
-      this.enableChildOverflow: false,
-      this.endRefresh,
-      this.onResetValue,
-      this.dismissType: BezierDismissType.RectSpread,
-      this.rectHeight: 70,
-      this.bezierColor})
-      : super(refreshStyle: RefreshStyle.UnFollow, height: rectHeight);
+  BezierHeader({
+    this.child = const Text(""),
+    this.onOffsetChange,
+    this.onModeChange,
+    this.readyRefresh,
+    this.enableChildOverflow = false,
+    this.endRefresh,
+    this.onResetValue,
+    this.dismissType = BezierDismissType.RectSpread,
+    this.rectHeight=  70,
+    this.bezierColor
+  }) : super(refreshStyle: RefreshStyle.UnFollow, height: rectHeight);
 
   @override
   State<StatefulWidget> createState() {
@@ -304,14 +299,15 @@ class BezierCircleHeader extends StatefulWidget {
 
   final BezierDismissType dismissType;
 
-  BezierCircleHeader(
-      {this.bezierColor,
-      this.rectHeight: 70,
-      this.circleColor: Colors.white,
-      this.enableChildOverflow: false,
-      this.dismissType: BezierDismissType.RectSpread,
-      this.circleType: BezierCircleType.Progress,
-      this.circleRadius: 12});
+  BezierCircleHeader({
+    this.bezierColor,
+    this.rectHeight = 70,
+    this.circleColor = Colors.white,
+    this.enableChildOverflow = false,
+    this.dismissType = BezierDismissType.RectSpread,
+    this.circleType = BezierCircleType.Progress,
+    this.circleRadius = 12
+  });
 
   @override
   State<StatefulWidget> createState() {
