@@ -4,12 +4,8 @@
  *   createTime:2018-05-14 17:39
  */
 
-import 'package:flutter/material.dart'
-    hide RefreshIndicator, RefreshIndicatorState;
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart' hide RefreshIndicator, RefreshIndicatorState;
 import '../../pull_to_refresh.dart';
-import '../internals/indicator_wrap.dart';
-import '../smart_refresher.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 
@@ -58,12 +54,12 @@ class ClassicHeader extends RefreshIndicator {
   final TextStyle textStyle;
 
   const ClassicHeader({
-    Key? key,
-    RefreshStyle refreshStyle: RefreshStyle.Follow,
-    double height: 60.0,
-    Duration completeDuration: const Duration(milliseconds: 600),
+    super.key,
+    RefreshStyle refreshStyle = RefreshStyle.Follow,
+    double height = 60.0,
+    Duration completeDuration = const Duration(milliseconds: 600),
     this.outerBuilder,
-    this.textStyle: const TextStyle(color: Colors.grey),
+    this.textStyle = const TextStyle(color: Colors.grey),
     this.releaseText,
     this.refreshingText,
     this.canTwoLevelIcon,
@@ -72,19 +68,18 @@ class ClassicHeader extends RefreshIndicator {
     this.completeText,
     this.failedText,
     this.idleText,
-    this.iconPos: IconPosition.left,
-    this.spacing: 15.0,
+    this.iconPos = IconPosition.left,
+    this.spacing = 15.0,
     this.refreshingIcon,
-    this.failedIcon: const Icon(Icons.error, color: Colors.grey),
-    this.completeIcon: const Icon(Icons.done, color: Colors.grey),
+    this.failedIcon = const Icon(Icons.error, color: Colors.grey),
+    this.completeIcon = const Icon(Icons.done, color: Colors.grey),
     this.idleIcon = const Icon(Icons.arrow_downward, color: Colors.grey),
     this.releaseIcon = const Icon(Icons.refresh, color: Colors.grey),
   }) : super(
-          key: key,
-          refreshStyle: refreshStyle,
-          completeDuration: completeDuration,
-          height: height,
-        );
+    refreshStyle: refreshStyle,
+    completeDuration: completeDuration,
+    height: height,
+  );
 
   @override
   State createState() {
@@ -215,31 +210,30 @@ class ClassicFooter extends LoadIndicator {
   final Duration completeDuration;
 
   const ClassicFooter({
-    Key? key,
+    super.key,
     VoidCallback? onClick,
-    LoadStyle loadStyle: LoadStyle.ShowAlways,
-    double height: 60.0,
+    LoadStyle loadStyle = LoadStyle.ShowAlways,
+    double height = 60.0,
     this.outerBuilder,
-    this.textStyle: const TextStyle(color: Colors.grey),
+    this.textStyle = const TextStyle(color: Colors.grey),
     this.loadingText,
     this.noDataText,
     this.noMoreIcon,
     this.idleText,
     this.failedText,
     this.canLoadingText,
-    this.failedIcon: const Icon(Icons.error, color: Colors.grey),
-    this.iconPos: IconPosition.left,
-    this.spacing: 15.0,
-    this.completeDuration: const Duration(milliseconds: 300),
+    this.failedIcon = const Icon(Icons.error, color: Colors.grey),
+    this.iconPos = IconPosition.left,
+    this.spacing = 15.0,
+    this.completeDuration = const Duration(milliseconds: 300),
     this.loadingIcon,
-    this.canLoadingIcon: const Icon(Icons.autorenew, color: Colors.grey),
+    this.canLoadingIcon = const Icon(Icons.autorenew, color: Colors.grey),
     this.idleIcon = const Icon(Icons.arrow_upward, color: Colors.grey),
   }) : super(
-          key: key,
-          loadStyle: loadStyle,
-          height: height,
-          onClick: onClick,
-        );
+    loadStyle: loadStyle,
+    height: height,
+    onClick: onClick,
+  );
 
   @override
   State<StatefulWidget> createState() {
