@@ -63,21 +63,21 @@ class CustomHeader extends RefreshIndicator {
   final VoidCallback? onResetValue;
 
   const CustomHeader({
-    Key? key,
+    super.key,
     required this.builder,
     this.readyToRefresh,
     this.endRefresh,
     this.onOffsetChange,
     this.onModeChange,
     this.onResetValue,
-    double height: 60.0,
-    Duration completeDuration: const Duration(milliseconds: 600),
-    RefreshStyle refreshStyle: RefreshStyle.Follow,
+    double height = 60.0,
+    Duration completeDuration = const Duration(milliseconds: 600),
+    RefreshStyle refreshStyle = RefreshStyle.Follow,
   }) : super(
-            key: key,
-            completeDuration: completeDuration,
-            refreshStyle: refreshStyle,
-            height: height);
+    completeDuration: completeDuration,
+    refreshStyle: refreshStyle,
+    height: height
+  );
 
   @override
   State<StatefulWidget> createState() {
@@ -147,19 +147,20 @@ class CustomFooter extends LoadIndicator {
 
   const CustomFooter({
     Key? key,
-    double height: 60.0,
+    double height = 60.0,
     this.onModeChange,
     this.onOffsetChange,
     this.readyLoading,
     this.endLoading,
-    LoadStyle loadStyle: LoadStyle.ShowAlways,
+    LoadStyle loadStyle = LoadStyle.ShowAlways,
     required this.builder,
     Function? onClick,
   }) : super(
-            key: key,
-            onClick: onClick as void Function()?,
-            loadStyle: loadStyle,
-            height: height);
+    key: key,
+    onClick: onClick as void Function()?,
+    loadStyle: loadStyle,
+    height: height
+  );
 
   @override
   State<StatefulWidget> createState() {
