@@ -4,10 +4,8 @@
  * Time: 2019/5/19 下午9:23
  */
 
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart'
-    hide RefreshIndicator, RefreshIndicatorState;
-import 'package:flutter/widgets.dart';
+
+import 'package:flutter/material.dart' hide RefreshIndicator, RefreshIndicatorState;
 import '../internals/indicator_wrap.dart';
 import '../smart_refresher.dart';
 
@@ -33,20 +31,19 @@ class MaterialClassicHeader extends RefreshIndicator {
   final Color? backgroundColor;
 
   const MaterialClassicHeader({
-    Key? key,
-    double height: 80.0,
+    super.key,
+    double height = 80.0,
     this.semanticsLabel,
     this.semanticsValue,
     this.color,
-    double offset: 0,
-    this.distance: 50.0,
+    double offset = 0,
+    this.distance = 50.0,
     this.backgroundColor,
   }) : super(
-          key: key,
-          refreshStyle: RefreshStyle.Front,
-          offset: offset,
-          height: height,
-        );
+    refreshStyle: RefreshStyle.Front,
+    offset: offset,
+    height: height,
+  );
 
   @override
   State<StatefulWidget> createState() {
@@ -95,7 +92,7 @@ class _MaterialClassicHeaderState
   @override
   void didUpdateWidget(covariant MaterialClassicHeader oldWidget) {
     // TODO: implement didUpdateWidget
-    _position = Scrollable.of(context)!.position;
+    _position = Scrollable.of(context).position;
     super.didUpdateWidget(oldWidget);
   }
 
@@ -113,7 +110,7 @@ class _MaterialClassicHeaderState
           alignment: Alignment.topCenter,
           child: RefreshProgressIndicator(
             semanticsLabel: widget.semanticsLabel ??
-                MaterialLocalizations?.of(context)
+                MaterialLocalizations.of(context)
                     .refreshIndicatorSemanticLabel,
             semanticsValue: widget.semanticsValue,
             value: floating ? null : _valueAni.value,
@@ -157,7 +154,7 @@ class _MaterialClassicHeaderState
   @override
   void didChangeDependencies() {
     final ThemeData theme = Theme.of(context);
-    _position = Scrollable.of(context)!.position;
+    _position = Scrollable.of(context).position;
     _valueColor = _positionController.drive(
       ColorTween(
         begin: (widget.color ?? theme.primaryColor).withOpacity(0.0),
@@ -193,22 +190,22 @@ class _MaterialClassicHeaderState
 /// attach the waterdrop effect to [MaterialClassicHeader]
 class WaterDropMaterialHeader extends MaterialClassicHeader {
   const WaterDropMaterialHeader({
-    Key? key,
+    super.key,
     String? semanticsLabel,
-    double distance: 60.0,
-    double offset: 0,
+    double distance = 60.0,
+    double offset = 0,
     String? semanticsValue,
-    Color color: Colors.white,
+    Color color = Colors.white,
     Color? backgroundColor,
   }) : super(
-            key: key,
-            height: 80.0,
-            color: color,
-            distance: distance,
-            offset: offset,
-            backgroundColor: backgroundColor,
-            semanticsValue: semanticsValue,
-            semanticsLabel: semanticsLabel);
+    height: 80.0,
+    color: color,
+    distance: distance,
+    offset: offset,
+    backgroundColor: backgroundColor,
+    semanticsValue: semanticsValue,
+    semanticsLabel: semanticsLabel
+  );
 
   @override
   State<StatefulWidget> createState() {
