@@ -5,14 +5,10 @@
  */
 
 import 'dart:async';
-import 'package:flutter/material.dart'
-    hide RefreshIndicatorState, RefreshIndicator;
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart' hide RefreshIndicatorState, RefreshIndicator;
 import 'package:flutter/foundation.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
-import '../internals/indicator_wrap.dart';
 import 'package:flutter/cupertino.dart';
-import '../smart_refresher.dart';
 
 /// QQ ios refresh  header effect
 class WaterDropHeader extends RefreshIndicator {
@@ -32,22 +28,22 @@ class WaterDropHeader extends RefreshIndicator {
   final Color waterDropColor;
 
   const WaterDropHeader({
-    Key? key,
+    super.key,
     this.refresh,
     this.complete,
-    Duration completeDuration: const Duration(milliseconds: 600),
+    Duration completeDuration = const Duration(milliseconds: 600),
     this.failed,
-    this.waterDropColor: Colors.grey,
-    this.idleIcon: const Icon(
+    this.waterDropColor = Colors.grey,
+    this.idleIcon = const Icon(
       Icons.autorenew,
       size: 15,
       color: Colors.white,
     ),
   }) : super(
-            key: key,
-            height: 60.0,
-            completeDuration: completeDuration,
-            refreshStyle: RefreshStyle.UnFollow);
+    height: 60.0,
+    completeDuration: completeDuration,
+    refreshStyle: RefreshStyle.UnFollow
+  );
 
   @override
   State<StatefulWidget> createState() {
@@ -165,17 +161,17 @@ class _WaterDropHeaderState extends RefreshIndicatorState<WaterDropHeader>
                     ),
                   ),
                   quarterTurns:
-                      Scrollable.of(context)!.axisDirection == AxisDirection.up
+                      Scrollable.of(context).axisDirection == AxisDirection.up
                           ? 10
                           : 0,
                 ),
                 Container(
                   alignment:
-                      Scrollable.of(context)!.axisDirection == AxisDirection.up
+                      Scrollable.of(context).axisDirection == AxisDirection.up
                           ? Alignment.bottomCenter
                           : Alignment.topCenter,
                   margin:
-                      Scrollable.of(context)!.axisDirection == AxisDirection.up
+                      Scrollable.of(context).axisDirection == AxisDirection.up
                           ? EdgeInsets.only(bottom: 12.0)
                           : EdgeInsets.only(top: 12.0),
                   child: widget.idleIcon,
