@@ -5,15 +5,10 @@
 */
 
 import 'package:flutter/gestures.dart';
-import 'package:flutter/physics.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter/foundation.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
 import 'package:pull_to_refresh/src/internals/slivers.dart';
-import 'internals/indicator_wrap.dart';
-import 'internals/refresh_physics.dart';
-import 'indicator/classic_indicator.dart';
-import 'indicator/material_indicator.dart';
 
 // ignore_for_file: INVALID_USE_OF_PROTECTED_MEMBER
 // ignore_for_file: INVALID_USE_OF_VISIBLE_FOR_TESTING_MEMBER
@@ -207,28 +202,27 @@ class SmartRefresher extends StatefulWidget {
   /// such as AnimatedList,RecordableList,doesn't allow to put into child,it will wrap it into SliverToBoxAdapter
   /// If you don't need pull down refresh ,just enablePullDown = false,
   /// If you  need pull up load ,just enablePullUp = true
-  SmartRefresher(
-      {Key? key,
-      required this.controller,
-      this.child,
-      this.header,
-      this.footer,
-      this.enablePullDown: true,
-      this.enablePullUp: false,
-      this.enableTwoLevel: false,
-      this.onRefresh,
-      this.onLoading,
-      this.onTwoLevel,
-      this.dragStartBehavior,
-      this.primary,
-      this.cacheExtent,
-      this.semanticChildCount,
-      this.reverse,
-      this.physics,
-      this.scrollDirection,
-      this.scrollController})
-      : builder = null,
-        super(key: key);
+  SmartRefresher({
+    super.key,
+    required this.controller,
+    this.child,
+    this.header,
+    this.footer,
+    this.enablePullDown = true,
+    this.enablePullUp = false,
+    this.enableTwoLevel = false,
+    this.onRefresh,
+    this.onLoading,
+    this.onTwoLevel,
+    this.dragStartBehavior,
+    this.primary,
+    this.cacheExtent,
+    this.semanticChildCount,
+    this.reverse,
+    this.physics,
+    this.scrollDirection,
+    this.scrollController
+  }) : builder = null;
 
   /// creates a widget help attach the refresh and load more function
   /// controller must not be null,builder must not be null
@@ -238,12 +232,12 @@ class SmartRefresher extends StatefulWidget {
   /// refresh above NestedScrollVIew,we must use this constrctor to implements refresh above NestedScrollView,but for now,NestedScrollView
   /// can not support overscroll out of edge
   SmartRefresher.builder({
-    Key? key,
+    super.key,
     required this.controller,
     required this.builder,
-    this.enablePullDown: true,
-    this.enablePullUp: false,
-    this.enableTwoLevel: false,
+    this.enablePullDown = true,
+    this.enablePullUp = false,
+    this.enableTwoLevel = false,
     this.onRefresh,
     this.onLoading,
     this.onTwoLevel,
@@ -257,8 +251,7 @@ class SmartRefresher extends StatefulWidget {
         semanticChildCount = null,
         dragStartBehavior = null,
         cacheExtent = null,
-        primary = null,
-        super(key: key);
+        primary = null;
 
   static SmartRefresher? of(BuildContext? context) {
     return context!.findAncestorWidgetOfExactType<SmartRefresher>();
@@ -424,36 +417,35 @@ class SmartRefresherState extends State<SmartRefresher> {
         dragStartBehavior: dragStartBehavior ?? DragStartBehavior.start,
         reverse: reverse ?? false,
       );
-    } else if (childView is Scrollable) {
-      body = Scrollable(
-        physics: _getScrollPhysics(
-            conf, childView.physics ?? AlwaysScrollableScrollPhysics()),
-        controller: childView.controller,
-        axisDirection: childView.axisDirection,
-        semanticChildCount: childView.semanticChildCount,
-        dragStartBehavior: childView.dragStartBehavior,
-        viewportBuilder: (context, offset) {
-          Viewport viewport =
-              childView.viewportBuilder(context, offset) as Viewport;
-          if (widget.enablePullDown) {
-            viewport.children.insert(
-                0,
-                widget.header ??
-                    (conf?.headerBuilder != null
-                        ? conf?.headerBuilder!()
-                        : null) ??
-                    defaultHeader);
-          }
-          //insert header or footer
-          if (widget.enablePullUp) {
-            viewport.children.add(widget.footer ??
-                (conf?.footerBuilder != null ? conf?.footerBuilder!() : null) ??
-                defaultFooter);
-          }
-          return viewport;
-        },
-      );
-    }
+    } else  body = Scrollable(
+      physics: _getScrollPhysics(
+          conf, childView.physics ?? AlwaysScrollableScrollPhysics()),
+      controller: childView.controller,
+      axisDirection: childView.axisDirection,
+      semanticChildCount: childView.semanticChildCount,
+      dragStartBehavior: childView.dragStartBehavior,
+      viewportBuilder: (context, offset) {
+        Viewport viewport =
+            childView.viewportBuilder(context, offset) as Viewport;
+        if (widget.enablePullDown) {
+          viewport.children.insert(
+              0,
+              widget.header ??
+                  (conf?.headerBuilder != null
+                      ? conf?.headerBuilder!()
+                      : null) ??
+                  defaultHeader);
+        }
+        //insert header or footer
+        if (widget.enablePullUp) {
+          viewport.children.add(widget.footer ??
+              (conf?.footerBuilder != null ? conf?.footerBuilder!() : null) ??
+              defaultFooter);
+        }
+        return viewport;
+      },
+    );
+  
     return body;
   }
 
@@ -595,12 +587,12 @@ class RefreshController {
   /// initialRefreshStatus: headerMode default value
   ///
   /// initialLoadStatus: footerMode default value
-  RefreshController(
-      {this.initialRefresh: false,
-      RefreshStatus? initialRefreshStatus,
-      LoadStatus? initialLoadStatus}) {
-    this.headerMode =
-        RefreshNotifier(initialRefreshStatus ?? RefreshStatus.idle);
+  RefreshController({
+    this.initialRefresh = false,
+    RefreshStatus? initialRefreshStatus,
+    LoadStatus? initialLoadStatus
+  }) {
+    this.headerMode = RefreshNotifier(initialRefreshStatus ?? RefreshStatus.idle);
     this.footerMode = RefreshNotifier(initialLoadStatus ?? LoadStatus.idle);
   }
 
@@ -650,11 +642,11 @@ class RefreshController {
   }
 
   /// make the header enter refreshing state,and callback onRefresh
-  Future<void>? requestRefresh(
-      {bool needMove: true,
-      bool needCallback: true,
-      Duration duration: const Duration(milliseconds: 500),
-      Curve curve: Curves.linear}) {
+  Future<void>? requestRefresh({
+    bool needMove = true,
+    bool needCallback = true,
+    Duration duration = const Duration(milliseconds: 500),
+    Curve curve = Curves.linear}) {
     assert(position != null,
         'Try not to call requestRefresh() before build,please call after the ui was rendered');
     if (isRefresh) return Future.value();
@@ -691,12 +683,14 @@ class RefreshController {
         headerMode!.value = RefreshStatus.refreshing;
       });
     }
+    return null;
   }
 
   /// make the header enter refreshing state,and callback onRefresh
-  Future<void> requestTwoLevel(
-      {Duration duration: const Duration(milliseconds: 300),
-      Curve curve: Curves.linear}) {
+  Future<void> requestTwoLevel({
+    Duration duration = const Duration(milliseconds: 300),
+    Curve curve = Curves.linear
+  }) {
     assert(position != null,
         'Try not to call requestRefresh() before build,please call after the ui was rendered');
     headerMode!.value = RefreshStatus.twoLevelOpening;
@@ -707,11 +701,12 @@ class RefreshController {
   }
 
   /// make the footer enter loading state,and callback onLoading
-  Future<void>? requestLoading(
-      {bool needMove: true,
-      bool needCallback: true,
-      Duration duration: const Duration(milliseconds: 300),
-      Curve curve: Curves.linear}) {
+  Future<void>? requestLoading({
+    bool needMove = true,
+      bool needCallback = true,
+      Duration duration = const Duration(milliseconds: 300),
+      Curve curve = Curves.linear
+    }) {
     assert(position != null,
         'Try not to call requestLoading() before build,please call after the ui was rendered');
     if (isLoading) return Future.value();
@@ -750,7 +745,7 @@ class RefreshController {
   /// request complete,the header will enter complete state,
   ///
   /// resetFooterState : it will set the footer state from noData to idle
-  void refreshCompleted({bool resetFooterState: false}) {
+  void refreshCompleted({bool resetFooterState = false}) {
     headerMode?.value = RefreshStatus.completed;
     if (resetFooterState) {
       resetNoData();
@@ -758,9 +753,9 @@ class RefreshController {
   }
 
   /// end twoLeveling,will return back first floor
-  Future<void>? twoLevelComplete(
-      {Duration duration: const Duration(milliseconds: 500),
-      Curve curve: Curves.linear}) {
+  Future<void>? twoLevelComplete({
+    Duration duration = const Duration(milliseconds: 500),
+    Curve curve = Curves.linear}) {
     headerMode?.value = RefreshStatus.twoLevelClosing;
     WidgetsBinding.instance.addPostFrameCallback((_) {
       position!
@@ -900,41 +895,41 @@ class RefreshConfiguration extends InheritedWidget {
   /// toggle of  loadmore vibrate
   final bool enableLoadMoreVibrate;
 
-  RefreshConfiguration(
-      {Key? key,
-      required this.child,
-      this.headerBuilder,
-      this.footerBuilder,
-      this.dragSpeedRatio: 1.0,
-      this.shouldFooterFollowWhenNotFull,
-      this.enableScrollWhenTwoLevel: true,
-      this.enableLoadingWhenNoData: false,
-      this.enableBallisticRefresh: false,
-      this.springDescription: const SpringDescription(
-        mass: 2.2,
-        stiffness: 150,
-        damping: 16,
-      ),
-      this.enableScrollWhenRefreshCompleted: false,
-      this.enableLoadingWhenFailed: true,
-      this.twiceTriggerDistance: 150.0,
-      this.closeTwoLevelDistance: 80.0,
-      this.skipCanRefresh: false,
-      this.maxOverScrollExtent,
-      this.enableBallisticLoad: true,
-      this.maxUnderScrollExtent,
-      this.headerTriggerDistance: 80.0,
-      this.footerTriggerDistance: 15.0,
-      this.hideFooterWhenNotFull: false,
-      this.enableRefreshVibrate: false,
-      this.enableLoadMoreVibrate: false,
-      this.topHitBoundary,
-      this.bottomHitBoundary})
-      : assert(headerTriggerDistance > 0),
-        assert(twiceTriggerDistance > 0),
-        assert(closeTwoLevelDistance > 0),
-        assert(dragSpeedRatio > 0),
-        super(key: key, child: child);
+  RefreshConfiguration({
+    super.key,
+    required this.child,
+    this.headerBuilder,
+    this.footerBuilder,
+    this.dragSpeedRatio = 1.0,
+    this.shouldFooterFollowWhenNotFull,
+    this.enableScrollWhenTwoLevel = true,
+    this.enableLoadingWhenNoData = false,
+    this.enableBallisticRefresh = false,
+    this.springDescription = const SpringDescription(
+      mass: 2.2,
+      stiffness: 150,
+      damping: 16,
+    ),
+    this.enableScrollWhenRefreshCompleted = false,
+    this.enableLoadingWhenFailed = true,
+    this.twiceTriggerDistance = 150.0,
+    this.closeTwoLevelDistance = 80.0,
+    this.skipCanRefresh = false,
+    this.maxOverScrollExtent,
+    this.enableBallisticLoad = true,
+    this.maxUnderScrollExtent,
+    this.headerTriggerDistance = 80.0,
+    this.footerTriggerDistance = 15.0,
+    this.hideFooterWhenNotFull = false,
+    this.enableRefreshVibrate = false,
+    this.enableLoadMoreVibrate = false,
+    this.topHitBoundary,
+    this.bottomHitBoundary})
+    : assert(headerTriggerDistance > 0),
+      assert(twiceTriggerDistance > 0),
+      assert(closeTwoLevelDistance > 0),
+      assert(dragSpeedRatio > 0),
+      super(child: child);
 
   /// Construct RefreshConfiguration to copy attributes from ancestor nodes
   /// If the parameter is null, it will automatically help you to absorb the attributes of your ancestor Refresh Configuration, instead of having to copy them manually by yourself.
